Rename scenes => pages
