+++ pages/Error.js
